==> notebooks/1-data_exploration.ipynb <==
import json

businesses = []
with open('../yelp_dataset/yelp_academic_dataset_business.json', 'r', encoding='utf-8') as f:
    for line in f:
        business = json.loads(line.strip())
        businesses.append(business)

businesses[0]

businesses[:11]

### Get businesses that are restaurants

restaurants = []
for b in businesses:
    if  b['categories'] is not None and ('food' in b['categories'].lower() or 'restaurant' in b['categories'].lower()):
        restaurants.append(b)
len(restaurants)


restaurants = {"restaurants": restaurants}
# with open('../yelp_dataset/yelp_restaurants.json', 'w', encoding='utf-8') as f:
#     json.dump(restaurants, f, indent=2)

rest_ids = set([r['business_id'] for r in restaurants['restaurants']])

category_list = []
for r in restaurants['restaurants']:
    categories = r['categories'].split(', ')
    category_list.extend(categories)
print(len(category_list))

len(set(category_list))

users = []
with open('../yelp_dataset/yelp_academic_dataset_user.json', 'r', encoding='utf-8') as f:
    for line in f:
        user = json.loads(line.strip())
        users.append(user)

len(users)

users[87]

reviews = []
with open('../yelp_dataset/yelp_academic_dataset_review.json', 'r', encoding='utf-8') as f:
    for line in f:
        review = json.loads(line.strip())
        reviews.append(review)

len(reviews)

reviews[5]

restaurant_reviews = [r for r in reviews if r['business_id'] in rest_ids]
len(restaurant_reviews)

with open('../yelp_dataset/yelp_restaurant_reviews.json', 'w', encoding='utf-8') as f:
    json.dump(restaurant_reviews, f, indent=2)

dates = [review['date'] for review in reviews]
max(dates)

min(dates)

checkins = []
with open('../yelp_dataset/yelp_academic_dataset_checkin.json', 'r', encoding='utf-8') as f:
    for line in f:
        checkin = json.loads(line.strip())
        checkins.append(checkin)

len(checkins)

checkins[0]

tips = []
with open('../yelp_dataset/yelp_academic_dataset_tip.json', 'r', encoding='utf-8') as f:
    for line in f:
        tip = json.loads(line.strip())
        tips.append(tip)

len(tips)

tips[0]
